Add basic modules for a half-functional programming language
This commit adds the following modules:

**Common**:
A header file where the #include operations for useful and shareable
modules will live.

**Chunk**:
The Solaria programming language will be built on top of the concept of
"chunks". Each chunk represents a set of bytecode instructions.

This initial module contains three functions
  - init_chunk (creates a new empty chunk).
  - write_chunk (write a byte into the chunk).
  - free_chunk (reclaim the memory for a chunk).

**Memory**:
This module aggregates all the memory utils Solaria will use.
It includes the macros
  - GROW_CAPACITY (self-explanatory).
  - GROW_ARRAY (wrapper for arrays with dynamic sizes).
  - FREE_ARRAY (self-explanatory).

And the function **reallocate**. This is the implementation of the
functionalities exposed by the macros "GROW_ARRAY" and "FREE_ARRAY".

**Debug**:
A module for facilitating inspection of bytecode. It includes the
functions:
  - disassemble_chunk (prints info about a whole chunk).
  - disassemble_instrunction (prints info about an instruction in a
    chunk).
  - simple_instruction (prints the instruction name and advances the
    offset to the next instruction).