Add OP_CONSTANT operation for storing constants
Our language needs to be able to store constants belonging to a chunk.
These constants will come from operations like "3 + 4", where each
literal ("3" and "4") is stored as a constant in the chunk struct.

In order to do so, this commit introduces two new files:
  - value.h
  - value.c

And the chunk struct also gets a new field called "constants" of type
`value_array_t`. This type is expressed in a similar way as to our chunk
struc:

```c
typedef struct value_array_t {
  int capacity;
  int count;
  value_t* values;
} value_array_t;
```

I.e., `value_array_t` is a dynamic-sized array, and therefore will need
all the functions to manage its memory and contents:
  - init_value_array
  - write_value_array
  - free_value_array
  - print_value (util function for debugging)

In order to make this functional in our under-development language, we
need to state that the chunk->code will store:

  * 1 byte for the instruction itself (OP_CONSTANT)
  * 1 byte for the index of the constant being loaded