Add support for keeping track of lines for error handling
Any serious programming language has a way to let users know the line
that invoked an error in their source code.

This first implementation uses a very basic (yet memory-inefficient)
encoding for lines. A new array of integers is stored as a field named
"lines" in the chunk struct (chunk_t). Each instruction of "code" is
paired with an item in the "lines" array. This makes it really easy to
pair codes with the line in which they happened.

A future implementation can use a different kind of encoding. An example
of a better encoding would be the "run-lengh encoding":
- https://en.wikipedia.org/wiki/Run-length_encoding