Add OP_CONSTANT_LONG operation code
Previously, Solaria had a limitation in which the OP_CONSTANT operation
code could only be linked with at most 256 different indexes of
constants. This is due to a limitation in the `uint8_t`.

This commit changes that by adding a `OP_CONSTANT_LONG` operation code.
The logic behind the creation of constants now goes as follows:
  1. If the number of constants in a chunk is < 256, associate the with
     the OP_CONSTANT op_code.
  2. Otherwise, associate them with an OP_CONSTANT_LONG op_code.