Change approach to store line numbers to use run-length encoding
The previous approach to storing line numbers was rather memory-naive.
It would store the line numbers sequentially for every op code, even if
they happened to be in the same line as the previous X other op codes,
thus generating a lot of duplication.

In order to optimise this memory issue, the new code uses a run-length
encoding approach. It only stores the line when the line number changes
from one op code to another, whilst keeping track of how many times a
particular line number persisted consecutively. This allows us to save
heaps of memory when multiple operations happen in the same line.

The only particular overhead is retrieving the line number. Some lookups
are necessary to get the line number, but since this only ever happens
during debug mode, or when the application crashes, this is a minimal
downside.