Add the virtual machine and the OP_NEGATE operation
The virtual machine is a stack-based vm. Our first implementation will
use a singleton pattern to manage the vm and its state. This is OK for
this prototype and it will facilitate not having to pass a pointer to
every function that uses the vm.

Another initial limitation of this VM is that its stack of values can
only hold at most 256 values. A future improvement is to grow that stack
dynamically as needed.

The constants are manipulated as a stack, and the OP_NEGATE function
basically pops the last element on this stack, and pushes it immediately
after with its negated value.