Add tests cases for arithmetic statements
This commits introduces 5 arithmetic instructions to test Solaria.
These instructions are simple equations that resolve to a particular
number.

They test, but also expose, the limitations of having a left-to-right
evaluation (when we don't have parenthesis). For example, the following
equations results in an unexpected value:

`1 + 2 * 3 = 9`

This will be fixed once we can parse the language better. At that stage,
we'll understand which operators take precedence over others.