Fixes the REPL code and spins up Pratt Parser
The REPL was broken, as the call to the interpreter was outsite the
expected loop.

The remaining additions are meant to service the `Pratt Parser` for the
language front end. Some unary and binary operations are serviced, so
that we can do basic arithmetics. The other operations aren't yet
finished.

TODO: Fix memory garbage in REPL mode when parsing lines. E.g try to
parse: (-1 + 2) * 3 - -4