Add support for string objects
In Solaria, strings are treated as objects.
The only key peculiarity (not even that peculiar) is that Solaria
accepts string concatenation using the "+" (plus) operator.

This means that this code is valid:
```solaria
"a" + "b" // "ab"
```

With strings, we've also introduced the concept of "object" types, and a
file `objects.c` to orchestrate objects' life spam.