Implement a hash table for Solaria
Solaria will use a hash table to initially store variables. The key will
be an object string, and the value can be anything.

This hash table is also embedded in the virtual machine so that it can
keep track of variables.

The main functionality (setting, getting, deleting) entries in the table
is provided in both table.h and table.c.